Struct parquet_format_safe::ColumnChunk

pub struct ColumnChunk {
    pub file_path: Option<String>,
    pub file_offset: i64,
    pub meta_data: Option<ColumnMetaData>,
    pub offset_index_offset: Option<i64>,
    pub offset_index_length: Option<i32>,
    pub column_index_offset: Option<i64>,
    pub column_index_length: Option<i32>,
    pub crypto_metadata: Option<ColumnCryptoMetaData>,
    pub encrypted_column_metadata: Option<Vec<u8>>,
}

Fields
file_path: Option<String>

File where column data is stored. If not set, assumed to be same file as metadata. This path is relative to the current file.

file_offset: i64

Byte offset in file_path to the ColumnMetaData *

meta_data: Option<ColumnMetaData>

Column metadata for this chunk. This is the same content as what is at file_path/file_offset. Having it here has it replicated in the file metadata.

offset_index_offset: Option<i64>

File offset of ColumnChunk’s OffsetIndex *

offset_index_length: Option<i32>

Size of ColumnChunk’s OffsetIndex, in bytes *

column_index_offset: Option<i64>

File offset of ColumnChunk’s ColumnIndex *

column_index_length: Option<i32>

Size of ColumnChunk’s ColumnIndex, in bytes *

crypto_metadata: Option<ColumnCryptoMetaData>

Crypto metadata of encrypted columns *

encrypted_column_metadata: Option<Vec<u8>>

Encrypted column metadata for this chunk *

Implementations

impl ColumnChunk

pub fn new<F1, F3, F4, F5, F6, F7, F8, F9>(
file_path: F1,
file_offset: i64,
meta_data: F3,
offset_index_offset: F4,
offset_index_length: F5,
column_index_offset: F6,
column_index_length: F7,
crypto_metadata: F8,
encrypted_column_metadata: F9
) -> ColumnChunkwhere
F1: Into<Option<String>>,
F3: Into<Option<ColumnMetaData>>,
F4: Into<Option<i64>>,
F5: Into<Option<i32>>,
F6: Into<Option<i64>>,
F7: Into<Option<i32>>,
F8: Into<Option<ColumnCryptoMetaData>>,
F9: Into<Option<Vec<u8>>>,

pub fn read_from_in_protocol<T: TInputProtocol>(
i_prot: &mut T
) -> Result<ColumnChunk>

pub async fn stream_from_in_protocol<T: TInputStreamProtocol>(
i_prot: &mut T
) -> Result<ColumnChunk>

pub fn write_to_out_protocol<T: TOutputProtocol>(
&self,
o_prot: &mut T
) -> Result<usize>

pub async fn write_to_out_stream_protocol<T: TOutputStreamProtocol>(
&self,
o_prot: &mut T
) -> Result<usize>