Struct parquet_format_safe::ConvertedType

pub struct ConvertedType(pub i32);

DEPRECATED: Common types used by frameworks(e.g. hive, pig) using parquet. ConvertedType is superseded by LogicalType. This enum should not be extended.

See LogicalTypes.md for conversion between ConvertedType and LogicalType.

Tuple Fields
0: i32
Implementations

impl ConvertedType

pub const UTF8: ConvertedType = _

a BYTE_ARRAY actually contains UTF8 encoded chars

pub const MAP: ConvertedType = _

a map is converted as an optional field containing a repeated key/value pair

pub const MAP_KEY_VALUE: ConvertedType = _

a key/value pair is converted into a group of two fields

pub const LIST: ConvertedType = _

a list is converted into an optional field containing a repeated field for its values

pub const ENUM: ConvertedType = _

an enum is converted into a binary field

pub const DECIMAL: ConvertedType = _

A decimal value.

This may be used to annotate binary or fixed primitive types. The underlying byte array stores the unscaled value encoded as two’s complement using big-endian byte order (the most significant byte is the zeroth element). The value of the decimal is the value * 10^{-scale}.

This must be accompanied by a (maximum) precision and a scale in the SchemaElement. The precision specifies the number of digits in the decimal and the scale stores the location of the decimal point. For example 1.23 would have precision 3 (3 total digits) and scale 2 (the decimal point is 2 digits over).

pub const DATE: ConvertedType = _

A Date

Stored as days since Unix epoch, encoded as the INT32 physical type.

pub const TIME_MILLIS: ConvertedType = _

A time

The total number of milliseconds since midnight. The value is stored as an INT32 physical type.

pub const TIME_MICROS: ConvertedType = _

A time.

The total number of microseconds since midnight. The value is stored as an INT64 physical type.

pub const TIMESTAMP_MILLIS: ConvertedType = _

A date/time combination

Date and time recorded as milliseconds since the Unix epoch. Recorded as a physical type of INT64.

pub const TIMESTAMP_MICROS: ConvertedType = _

A date/time combination

Date and time recorded as microseconds since the Unix epoch. The value is stored as an INT64 physical type.

pub const UINT_8: ConvertedType = _

An unsigned integer value.

The number describes the maximum number of meaningful data bits in the stored value. 8, 16 and 32 bit values are stored using the INT32 physical type. 64 bit values are stored using the INT64 physical type.

pub const UINT_16: ConvertedType = _

pub const UINT_32: ConvertedType = _

pub const UINT_64: ConvertedType = _

pub const INT_8: ConvertedType = _

A signed integer value.

The number describes the maximum number of meaningful data bits in the stored value. 8, 16 and 32 bit values are stored using the INT32 physical type. 64 bit values are stored using the INT64 physical type.

pub const INT_16: ConvertedType = _

pub const INT_32: ConvertedType = _

pub const INT_64: ConvertedType = _

pub const JSON: ConvertedType = _

An embedded JSON document

A JSON document embedded within a single UTF8 column.

pub const BSON: ConvertedType = _

An embedded BSON document

A BSON document embedded within a single BINARY column.

pub const INTERVAL: ConvertedType = _

An interval of time

This type annotates data stored as a FIXED_LEN_BYTE_ARRAY of length 12 This data is composed of three separate little endian unsigned integers. Each stores a component of a duration of time. The first integer identifies the number of months associated with the duration, the second identifies the number of days associated with the duration and the third identifies the number of milliseconds associated with the provided duration. This duration of time is independent of any particular timezone or date.

pub const ENUM_VALUES: &'static [Self] = _

pub fn write_to_out_protocol<T: TOutputProtocol>(
&self,
o_prot: &mut T
) -> Result<usize>

pub async fn write_to_out_stream_protocol<T: TOutputStreamProtocol>(
&self,
o_prot: &mut T
) -> Result<usize>