Trait parquet_format_safe::thrift::protocol::TInputStreamProtocol

pub trait TInputStreamProtocol: Send + Sized {

    fn read_message_begin<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<TMessageIdentifier>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_message_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_struct_begin<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Option<TStructIdentifier>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_struct_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_field_begin<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<TFieldIdentifier>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_field_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_bool<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_bytes<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_i8<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<i8>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_i16<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<i16>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_i32<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<i32>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_i64<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<i64>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_double<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<f64>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_string<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_list_begin<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<TListIdentifier>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_list_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_set_begin<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<TSetIdentifier>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_set_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_map_begin<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<TMapIdentifier>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_map_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_byte<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<u8>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn read_list<'life0, 'async_trait, P>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<Vec<P>>> + Send + 'async_trait>>
    where
        P: 'async_trait + AsyncReadThrift,
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn skip<'life0, 'async_trait>(
        &'life0 mut self,
        field_type: TType
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }
    fn skip_till_depth<'life0, 'async_trait>(
        &'life0 mut self,
        field_type: TType,
        depth: i8
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait,
    { ... }

}

Required Methods

fn read_message_begin<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<TMessageIdentifier>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the beginning of a Thrift message.

fn read_message_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the end of a Thrift message.

fn read_struct_begin<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Option<TStructIdentifier>>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the beginning of a Thrift struct.

fn read_struct_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the end of a Thrift struct.

fn read_field_begin<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<TFieldIdentifier>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the beginning of a Thrift struct field.

fn read_field_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the end of a Thrift struct field.

fn read_bool<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<bool>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read a bool.

fn read_bytes<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Vec<u8>>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read a fixed-length byte array.

fn read_i8<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<i8>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read a word.

fn read_i16<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<i16>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read a 16-bit signed integer.

fn read_i32<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<i32>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read a 32-bit signed integer.

fn read_i64<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<i64>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read a 64-bit signed integer.

fn read_double<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<f64>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read a 64-bit float.

fn read_string<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read a fixed-length string (not null terminated).

fn read_list_begin<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<TListIdentifier>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the beginning of a list.

fn read_list_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the end of a list.

fn read_set_begin<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<TSetIdentifier>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the beginning of a set.

fn read_set_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the end of a set.

fn read_map_begin<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<TMapIdentifier>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the beginning of a map.

fn read_map_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read the end of a map.

fn read_byte<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<u8>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Read an unsigned byte.

This method should never be used in generated code.

Provided Methods

fn read_list<'life0, 'async_trait, P>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<Vec<P>>> + Send + 'async_trait>>where
P: 'async_trait + AsyncReadThrift,
'life0: 'async_trait,
Self: 'async_trait,

fn skip<'life0, 'async_trait>(
&'life0 mut self,
field_type: TType
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Skip a field with type field_type recursively until the default maximum skip depth is reached.

fn skip_till_depth<'life0, 'async_trait>(
&'life0 mut self,
field_type: TType,
depth: i8
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Skip a field with type field_type recursively up to depth levels.