Trait parquet_format_safe::thrift::protocol::TOutputStreamProtocol

pub trait TOutputStreamProtocol: Send {

    fn write_message_begin<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        identifier: &'life1 TMessageIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn write_message_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_struct_begin<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        identifier: &'life1 TStructIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn write_struct_end(&mut self) -> Result<usize>;
    fn write_field_begin<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        identifier: &'life1 TFieldIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn write_field_end(&mut self) -> Result<usize>;
    fn write_field_stop<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_bool<'life0, 'async_trait>(
        &'life0 mut self,
        b: bool
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_bytes<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        b: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn write_i8<'life0, 'async_trait>(
        &'life0 mut self,
        i: i8
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_i16<'life0, 'async_trait>(
        &'life0 mut self,
        i: i16
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_i32<'life0, 'async_trait>(
        &'life0 mut self,
        i: i32
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_i64<'life0, 'async_trait>(
        &'life0 mut self,
        i: i64
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_double<'life0, 'async_trait>(
        &'life0 mut self,
        d: f64
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_string<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        s: &'life1 str
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn write_list_begin<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        identifier: &'life1 TListIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn write_list_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_set_begin<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        identifier: &'life1 TSetIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn write_set_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_map_begin<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        identifier: &'life1 TMapIdentifier
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn write_map_end<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn flush<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn write_byte<'life0, 'async_trait>(
        &'life0 mut self,
        b: u8
    ) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

}

Required Methods

fn write_message_begin<'life0, 'life1, 'async_trait>(
&'life0 mut self,
identifier: &'life1 TMessageIdentifier
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Write the beginning of a Thrift message.

fn write_message_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write the end of a Thrift message.

fn write_struct_begin<'life0, 'life1, 'async_trait>(
&'life0 mut self,
identifier: &'life1 TStructIdentifier
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Write the beginning of a Thrift struct.

fn write_struct_end(&mut self) -> Result<usize>

Write the end of a Thrift struct.

fn write_field_begin<'life0, 'life1, 'async_trait>(
&'life0 mut self,
identifier: &'life1 TFieldIdentifier
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Write the beginning of a Thrift field.

fn write_field_end(&mut self) -> Result<usize>

Write the end of a Thrift field.

fn write_field_stop<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write a STOP field indicating that all the fields in a struct have been written.

fn write_bool<'life0, 'async_trait>(
&'life0 mut self,
b: bool
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write a bool.

fn write_bytes<'life0, 'life1, 'async_trait>(
&'life0 mut self,
b: &'life1 [u8]
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Write a fixed-length byte array.

fn write_i8<'life0, 'async_trait>(
&'life0 mut self,
i: i8
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write an 8-bit signed integer.

fn write_i16<'life0, 'async_trait>(
&'life0 mut self,
i: i16
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write a 16-bit signed integer.

fn write_i32<'life0, 'async_trait>(
&'life0 mut self,
i: i32
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write a 32-bit signed integer.

fn write_i64<'life0, 'async_trait>(
&'life0 mut self,
i: i64
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write a 64-bit signed integer.

fn write_double<'life0, 'async_trait>(
&'life0 mut self,
d: f64
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write a 64-bit float.

fn write_string<'life0, 'life1, 'async_trait>(
&'life0 mut self,
s: &'life1 str
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Write a fixed-length string.

fn write_list_begin<'life0, 'life1, 'async_trait>(
&'life0 mut self,
identifier: &'life1 TListIdentifier
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Write the beginning of a list.

fn write_list_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write the end of a list.

fn write_set_begin<'life0, 'life1, 'async_trait>(
&'life0 mut self,
identifier: &'life1 TSetIdentifier
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Write the beginning of a set.

fn write_set_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write the end of a set.

fn write_map_begin<'life0, 'life1, 'async_trait>(
&'life0 mut self,
identifier: &'life1 TMapIdentifier
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Write the beginning of a map.

fn write_map_end<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write the end of a map.

fn flush<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Flush buffered bytes to the underlying transport.

fn write_byte<'life0, 'async_trait>(
&'life0 mut self,
b: u8
) -> Pin<Box<dyn Future<Output = Result<usize>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Write an unsigned byte.

This method should never be used in generated code.